chipod depth when calculating mooring velocity
I often have many depths in my moor structure. Changed the way the
input arguments are defined in pre-driver when it calls
chi_generate_vel_adcp. moor.depth can have many depths and ChipodDepth
is the depth of the chipod that will be processed. This should not
cause any problems with Deepak’s ExtractTSFromTaoTritonPirataRama