removed the capabilities in pre_driver to do_temp and do_dTdz_i
In pre_driver.m, there are two flags:
- do_temp
- do_dTdz_i
which just call other functions:
- do_temp_proc
- do_dTdz_i_proc
BUT at the beginning of do_temp_proc and do_dTdz_i proc there are flags
at the beginning that overwrite the flags that the user THINKS they are
setting in pre_driver which cause the code to crash.

Users should run do_temp_proc and do_dTdz_i_proc independently and NOT
from pre_driver.