Set chi, eps to 0 if Tp is below noise floor
Estimate spectral noise floor, if below that set chi to 0. Else nan if fit
fails.
I do this by comparing variance to noise floor variance.
Tolerance of factor of 4 in variance seems to catch most noisy sections in Tp.
This is a good improvement because even if we miss some, we are not
systematically setting chi = NaN when there is very little variance