Massive speed improvement in chi_merge_and_avg.m
Previously, we appended to an ever-increasing array. This meant that MATLAB
repeatedly copied and extended the size of the array, slowing it down a lot.

Now we loop through files append to a pre-allocated cell array. And then merge
things together by taking advantage of MATLAB's builtin concatenation of struct
array fields. This is waaaaaaaaay faster.