Added additional datenum format in whoAmI_timelimits.m
ALL of the code has been changed so I have to define time limits in the
sqlite database rather than being able to easily define these
parameters in the mfiles. It would have been nice if the code had been
left so that EITHER the times could be read from the sqlite database OR
defined in the code. That being said, I tend to use datenum in the
format ‘yyyy,mm,dd,HH,MM,SS’, so I needed to add that additional format
to whoAmI_timelimits.m