update default settings for calibrate_pitot.m
The comment says, "averaging window for praw (default 600 [sec])," therefore the default should ACTUALLY BE 600 seconds not 30 seconds as it was set. I changed some other defaults settings as well.

These are not things that should be changed in the master copy of the code for a specific deployment because there are many users who rely on this code.