no N2 when pressure sensor dies
When calculating N2 with do_dTdz_i.m, the code calls cal_N2_from_TS.m to calculate N2. If the pressure sensor dies and is NaN, the calculated values for alpha and beta also end up being NaNs which results in N2 being NaN even if there are good values for dTdz. There is almost no dependence on pressure for alpha and beta, so I've just set the pressure input for this calculation to be 50dbar if pressure is NaN.