Update compare_dTdz.m to have good time limits for the plot
When processing do_dTdz_i_proc, the plot that is made uses the time limits set in whoAmI.m. However, if the time limits have not already been set, the default goes from 2000-2050, which makes it impossible to see the data. Adding an if-statement to override the default time limits with the start and end time of the data in cases where the time limits have not been set in whoAmI.