Ensure that Kt and Jq are not accidentely all NaN in Turb.mat
When the depth is not defined in whoAmI.m, it results in Kt and Jq accidentely being all NaN in Turb.mat. So I decided to add a fix to do_combine_turb.m.

The difference between sw_tdif (molecular diffusivity) at 0m vs 100m is ~0.03% and is likely MUCH smaller than the turbulence diffusivity anyway, so this has no effect on the final result.

Pavan had multiple chipods where he encountered this issue, so a work-around was needed in do_combine_turb.m. Obviously, users should set the depth in whoAmI.m, but if they don't, it shouldn't cause an error of this magnitude.